allow to sign messages for Terra
