+++ src/ActionModal/utils/MessageConstructor.js
@@ -14,12 +14,9 @@ export const getMessage = async (
 
 const getMessageConstructor = async context => {
   switch (context.networkId) {
-    case `local-cosmos-hub-testnet`: {
-      const { default: Cosmos } = await import("cosmos-apiV0")
-      const cosmos = new Cosmos(context.url || "", context.chainId || "")
-      return (messageType, userAddress, transactionProperties) =>
-        cosmos[messageType](userAddress, transactionProperties)
-    }
+    case `local-cosmos-hub-testnet`:
+    case `terra-mainnet`:
+    case `terra-testnet`:
     case `cosmos-hub-mainnet`:
     case `cosmos-hub-testnet`: {
       const { default: Cosmos } = await import("cosmos-apiV2")
@@ -33,10 +30,9 @@ const getMessageConstructor = async context => {
 
 export const getTransactionSigner = async context => {
   switch (context.networkId) {
-    case `local-cosmos-hub-testnet`: {
-      const { createSignedTransaction } = await import("cosmos-apiV0")
-      return createSignedTransaction
-    }
+    case `local-cosmos-hub-testnet`:
+    case `terra-mainnet`:
+    case `terra-testnet`:
     case `cosmos-hub-mainnet`:
     case `cosmos-hub-testnet`: {
       const { createSignedTransaction } = await import("cosmos-apiV2")
@@ -48,11 +44,9 @@ export const getTransactionSigner = async context => {
 
 export const getMultiMessage = async (context, messages) => {
   switch (context.networkId) {
-    case `local-cosmos-hub-testnet`: {
-      const { default: Cosmos } = await import("cosmos-apiV0")
-      const cosmos = new Cosmos(context.url || "", context.chainId || "")
-      return cosmos.MultiMessage(context.userAddress, messages)
-    }
+    case `local-cosmos-hub-testnet`:
+    case `terra-mainnet`:
+    case `terra-testnet`:
     case `cosmos-hub-mainnet`:
     case `cosmos-hub-testnet`: {
       const { default: Cosmos } = await import("cosmos-apiV2")
@@ -62,174 +56,31 @@ export const getMultiMessage = async (context, messages) => {
   }
 }
 
-// Bank
 /* istanbul ignore next */
-export function MsgSend(
+async function getMessageFormatter(network, messageType) {
+  let networkMessages
+  try {
+    networkMessages = await import(`./networkMessages/${network}.js`)
+  } catch (err) {
+    throw new Error("Signing for this network is not enabled.")
+  }
+  const messageFormatter = networkMessages[messageType]
+
+  if (!messageFormatter) {
+    throw new Error(
+      `Creating the message ${messageType} for the network ${network} is not supported`
+    )
+  }
+
+  return messageFormatter
+}
+
+export async function transformMessage(
+  network,
+  messageType,
   senderAddress,
-  {
-    toAddress,
-    amounts // [{ denom, amount}]
-  }
+  message
 ) {
-  return {
-    type: `cosmos-sdk/MsgSend`,
-    value: {
-      from_address: senderAddress,
-      to_address: toAddress,
-      amount: amounts.map(Coin)
-    }
-  }
-}
-
-// Staking
-export function MsgDelegate(
-  senderAddress,
-  { validatorAddress, amount, denom }
-) {
-  /* istanbul ignore next */
-  return {
-    type: `cosmos-sdk/MsgDelegate`,
-    value: {
-      delegator_address: senderAddress,
-      validator_address: validatorAddress,
-      amount: Coin({ amount, denom })
-    }
-  }
-}
-
-export function MsgUndelegate(
-  senderAddress,
-  { validatorAddress, amount, denom }
-) {
-  /* istanbul ignore next */
-  return {
-    type: `cosmos-sdk/MsgUndelegate`,
-    value: {
-      validator_address: validatorAddress,
-      delegator_address: senderAddress,
-      amount: Coin({ amount, denom })
-    }
-  }
-}
-
-export function MsgRedelegate(
-  senderAddress,
-  { validatorSourceAddress, validatorDestinationAddress, amount, denom }
-) {
-  /* istanbul ignore next */
-  return {
-    type: `cosmos-sdk/MsgBeginRedelegate`,
-    value: {
-      delegator_address: senderAddress,
-      validator_src_address: validatorSourceAddress,
-      validator_dst_address: validatorDestinationAddress,
-      amount: Coin({ amount, denom })
-    }
-  }
-}
-
-// Governance
-export function MsgSubmitProposal(
-  senderAddress,
-  {
-    title,
-    description,
-    initialDeposits // [{ denom, amount }]
-  }
-) {
-  /* istanbul ignore next */
-  return {
-    type: `cosmos-sdk/MsgSubmitProposal`,
-    value: {
-      content: {
-        type: "cosmos-sdk/TextProposal",
-        value: {
-          title,
-          description
-        }
-      },
-      proposer: senderAddress,
-      initial_deposit: initialDeposits.map(Coin)
-    }
-  }
-}
-
-export function MsgVote(senderAddress, { proposalId, option }) {
-  /* istanbul ignore next */
-  return {
-    type: `cosmos-sdk/MsgVote`,
-    value: {
-      voter: senderAddress,
-      proposal_id: proposalId,
-      option
-    }
-  }
-}
-
-export function MsgDeposit(
-  senderAddress,
-  {
-    proposalId,
-    amounts // [{ denom, amount }]
-  }
-) {
-  /* istanbul ignore next */
-  return {
-    type: `cosmos-sdk/MsgDeposit`,
-    value: {
-      depositor: senderAddress,
-      proposal_id: proposalId,
-      amount: amounts.map(Coin)
-    }
-  }
-}
-
-export function MsgWithdrawDelegationReward(
-  senderAddress,
-  { validatorAddress }
-) {
-  /* istanbul ignore next */
-  return {
-    type: `cosmos-sdk/MsgWithdrawDelegationReward`,
-    value: {
-      delegator_address: senderAddress,
-      validator_address: validatorAddress
-    }
-  }
-}
-
-function Coin({ amount, denom }) {
-  return {
-    amount: String(amount),
-    denom
-  }
-}
-
-/* istanbul ignore next */
-function getMessageFormatter(messageType) {
-  switch (messageType) {
-    case "MsgSend":
-      return MsgSend
-    case "MsgDelegate":
-      return MsgDelegate
-    case "MsgUndelegate":
-      return MsgUndelegate
-    case "MsgRedelegate":
-      return MsgRedelegate
-    case "MsgSubmitProposal":
-      return MsgSubmitProposal
-    case "MsgVote":
-      return MsgVote
-    case "MsgDeposit":
-      return MsgDeposit
-    case "MsgWithdrawDelegationReward":
-      return MsgWithdrawDelegationReward
-    default:
-      return null
-  }
-}
-
-export function transformMessage(messageType, senderAddress, message) {
-  const messageFormatter = getMessageFormatter(messageType)
+  const messageFormatter = await getMessageFormatter(network, messageType)
   return messageFormatter(senderAddress, message)
 }
