+++ src/ActionModal/utils/networkMessages/cosmos-hub-testnet.js
@@ -1,0 +1,1 @@
+export * from "./cosmos-hub-mainnet.js"
