+++ src/ActionModal/utils/networkMessages/terra-testnet.js
@@ -1,0 +1,9 @@
+import * as CosmosMessages from "./cosmos-hub-mainnet.js"
+
+export const MsgSend = (...args) => {
+  const msg = CosmosMessages.MsgSend(...args)
+  msg.type = "bank/MsgSend"
+  return msg
+}
+
+// TODO all message types
